output float64

+++ tools/get_texture.py
@@ -169,7 +169,8 @@ def main():
 
     if args.verbose:
         print('Writing %s features to %s' % (len(feats), args.output))
-    feats = np.array([feats], dtype=np.float32)
+    #feats = np.array([feats], dtype=np.float32)
+    feats = np.array([feats])
     #dwi.asciifile.write_ascii_file(args.output, feats, featnames)
     dwi.files.write_pmap(args.output, feats, featnames)
